chore: runner class seperate for testing feat: cli arguments parse by `argparse` feat: user_agent move into praw.ini
