+++ .gitignore
@@ -202,3 +202,4 @@ $RECYCLE.BIN/
 # Custom rules (everything added below won't be overriden by 'Generate .gitignore File' if you use 'Update' option)
 *.ini
 !praw.example.ini
+.vscode/settings.json
