+++ praw.example.ini
@@ -3,3 +3,4 @@ username      = <username>
 password      = <password>
 client_id     = <client_id>
 client_secret = <client_secret>
+user_agent    = User-Agent: linux:com.%(username)s.runner:v1.0 (by u/%(username)s)
